+++ entities/ticket_note.go
@@ -12,8 +12,8 @@ type TicketNote struct {
 	TicketID          autotask.Optional[int64]     `json:"ticketID,omitzero"`
 	Title             autotask.Optional[string]    `json:"title,omitzero"`
 	Description       autotask.Optional[string]    `json:"description,omitzero"`
-	NoteType          autotask.Optional[int]       `json:"noteType,omitzero"`
-	Publish           autotask.Optional[int]       `json:"publish,omitzero"`
+	NoteType          autotask.Optional[int64]       `json:"noteType,omitzero"`
+	Publish           autotask.Optional[int64]       `json:"publish,omitzero"`
 	CreateDateTime    autotask.Optional[time.Time] `json:"createDateTime,omitzero"`
 	CreatorResourceID autotask.Optional[int64]     `json:"creatorResourceID,omitzero"`
 	LastActivityDate  autotask.Optional[time.Time] `json:"lastActivityDate,omitzero"`
